% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_naive.R
\name{simple4}
\alias{simple4}
\title{A simple4 function}
\usage{
simple4(x, y, z, aa)
}
\arguments{
\item{x}{First}

\item{y}{Second}

\item{z}{Third}

\item{aa}{Fourth}
}
\description{
A simple4 function
}
